change main API to createAction with Form property
