+++ packages/start/form/index.tsx
@@ -1,5 +1,5 @@
-export { Form, createForm, FormError } from "./Form";
+export { Form, FormError } from "./Form";
 export type { FormSubmission, SubmitOptions, FormAction, FormProps, FormMethod } from "./Form";
 
-export { createAction } from "./action";
-export type { ActionSubmission } from "./action";
+export { createActionState, createAction } from "./createAction";
+export type { ActionSubmission, Action } from "./createAction";
